migrationnya di bujuri user udah add datatable

+++ database/migrations/2018_04_13_121345_create_Users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration {
 			$table->string('username', 20)->unique();
 			$table->string('password', 255);
 			$table->string('Telepon', 13);
-			$table->float('gaji');
+			$table->double('gaji');
 			$table->text('Alamat');
 			$table->integer('user_id')->unsigned()->nullable();
 		});
